move keyboard & control_abbreviations to headless
